+++ src/main/java/lotto/model/LottoGenerator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-import static lotto.Winning.*;
+import static lotto.model.Winning.*;
 
 public class LottoGenerator {
     private final LottoCollection collection;
