+++ src/main/java/lotto/view/ResultView.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.view;
+
+import lotto.model.LottoCollection;
 
 import java.util.List;
