+++ src/test/java/lotto/WinningTest.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.model.LottoCollection;
+import lotto.model.Winning;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
